pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finlit-agent"
version = "0.1.0"
description = "A financial literacy agent for personal finance management."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Financial Literacy Agent Team", email = "contact@example.com"},
]
license = {text = "MIT"}
keywords = ["ai", "agent", "financial", "literacy", "ollama", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "accelerate>=1.10.0",
    "smolagents[litellm,mcp,toolkit]>=1.21.1",
    "torch>=2.8.0",
    "transformers>=4.55.0",
    "litellm>=1.55.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/christoschr97/greek-finlit-agent"
Repository = "https://github.com/christoschr97/greek-finlit-agent"
Documentation = "https://github.com/christoschr97/greek-finlit-agent#readme"
Issues = "https://github.com/christoschr97/greek-finlit-agent/issues"

[project.scripts]
finlit-agent = "main:main"
finlit-models = "scripts.models:main"
finlit-quickstart = "scripts.quickstart:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.991",
]
